Added functions to ArchiveClient, and ArchiveModule to obtain archive information from query values.
